requirements.txt
paho-mqtt
adafruit-circuitpython-tlc5947